refactor: organize crypto utils
