+++ frontend/utils/crypto/constants.ts
@@ -1,0 +1,1 @@
+export const VERSION = 1
